TomoPy 0.1.14

+++ tomopy/misc/morph.py
@@ -50,7 +50,10 @@
 Module for data size morphing functions.
 """
 
-from __future__ import absolute_import, division, print_function
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
+from __future__ import unicode_literals
 
 import numpy as np
 import tomopy.util.extern as extern
@@ -184,9 +187,9 @@ def _sample(arr, level, axis, mode):
     dx, dy, dz = arr.shape
 
     if mode == 0:
-        dim = arr.shape[axis] / np.power(2, level)
+        dim = int(arr.shape[axis] / np.power(2, level))
     if mode == 1:
-        dim = arr.shape[axis] * np.power(2, level)
+        dim = int(arr.shape[axis] * np.power(2, level))
 
     out = _init_out(arr, axis, dim)
     return extern.c_sample(mode, arr, dx, dy, dz, level, axis, out)
